refactor file placement
