+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/CMakeLists.txt
@@ -10,7 +10,7 @@ rj_add_object_library(rocjitsu_code
     relocation_function_table.cpp
     rj_code.cpp
     dbt/binary_translator.cpp
-    dbt/gfx1250_b0_to_a0.cpp
+    dbt/legalization/gfx1250_b0_to_a0.cpp
     dbt/rj_code_translate.cpp
     dbt/hazard_tracker.cpp
     dbt/kernel_descriptor_translator.cpp
@@ -23,6 +23,7 @@ rj_add_object_library(rocjitsu_code
     dbt/semantic/cdna4_to_rdna_common.cpp
     dbt/semantic/cdna4_to_rdna3.cpp
     dbt/semantic/cdna4_to_rdna4.cpp
+    dbt/semantic/gfx1250_b0_to_a0.cpp
     dbt/semantic_translator.cpp
     dbt/semantic_scratch.cpp
     dbt/virtual_lds.cpp
