+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/dbt/legalization/gfx1250_b0_to_a0.h
@@ -4,8 +4,8 @@
 /// @file gfx1250_b0_to_a0.h
 /// @brief gfx1250 B0-to-A0 errata legalization classification.
 
-#ifndef ROCJITSU_CODE_DBT_GFX1250_B0_TO_A0_H_
-#define ROCJITSU_CODE_DBT_GFX1250_B0_TO_A0_H_
+#ifndef ROCJITSU_CODE_DBT_LEGALIZATION_GFX1250_B0_TO_A0_H_
+#define ROCJITSU_CODE_DBT_LEGALIZATION_GFX1250_B0_TO_A0_H_
 
 namespace rocjitsu {
 
@@ -29,4 +29,4 @@ struct InstructionLegalization;
 
 } // namespace rocjitsu
 
-#endif // ROCJITSU_CODE_DBT_GFX1250_B0_TO_A0_H_
+#endif // ROCJITSU_CODE_DBT_LEGALIZATION_GFX1250_B0_TO_A0_H_
